Relocated use case folders to relative ui folder.
