+++ src/commonMain/kotlin/onlytrade/app/viewmodel/home/di/HomeDI.kt
@@ -1,7 +1,7 @@
 package onlytrade.app.viewmodel.home.di
 
 import onlytrade.app.viewmodel.home.ui.HomeViewModel
-import onlytrade.app.viewmodel.home.usecase.GetProductsUseCase
+import onlytrade.app.viewmodel.home.ui.usecase.GetProductsUseCase
 import onlytrade.app.viewmodel.product.repository.ProductRepository
 import onlytrade.app.viewmodel.product.repository.data.remote.api.GetProductsApi
 import org.koin.core.module.dsl.factoryOf
