+++ src/commonMain/kotlin/onlytrade/app/viewmodel/login/di/LoginDI.kt
@@ -3,8 +3,8 @@ package onlytrade.app.viewmodel.login.di
 import onlytrade.app.viewmodel.login.repository.LoginRepository
 import onlytrade.app.viewmodel.login.repository.data.remote.LoginApi
 import onlytrade.app.viewmodel.login.ui.LoginViewModel
-import onlytrade.app.viewmodel.login.usecase.EmailLoginUseCase
-import onlytrade.app.viewmodel.login.usecase.PhoneLoginUseCase
+import onlytrade.app.viewmodel.login.ui.usecase.EmailLoginUseCase
+import onlytrade.app.viewmodel.login.ui.usecase.PhoneLoginUseCase
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
